# code-generation-demo/src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Sample API
  description: API for demonstrating API-first approach with Spring Boot
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /greeting:
    get:
      tags:
        - EmployeeAPI
      summary: Returns a greeting message
      operationId: getGreeting
      responses:
        '200':
          description: A greeting message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Hello, World!
  /greeting1:
    get:
      tags:
        - EmployeeAPI
      summary: Returns a greeting message
      operationId: getGreeting1
      responses:
        '200':
          description: A greeting message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Hello, World!
  /employees:
    post:
      tags:
        - EmployeeAPI
      summary: Add a new employee
      operationId: addEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 1
        department:
          type: string
          minLength: 1
      required:
        - name
        - age
        - department
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        age:
          type: integer
        department:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string